backend/requirements.txt
Flask
gunicorn
requests
pytest-cov